Aiming For Targeted Hypotheses

The aim function does as described, and aims the bullet at the appropriate target. It allows for additional specification of how the hypothesis should be tested. It arranges these hypothesis into a table that outlines the trajectory of that bullet. Each part of the hypothesis ends up being a single row. It creates model identifiers, and collects both the pre-analysis inputs and the post-analysis attributes. It allows for organizing several related hypotheses together.

aim(bullets, ...)

Arguments

...
For extensibility

`bullets`
A list of bullet object(s) that will be used to generate a specific aim. Several bullet objects can be given in the list. The object names should be informative for the user. Each bullet should be from the same dataset to be able to run the analysis in the pre-specified models.

Value

An aim object that contain all the relevant hypothesis that were pre-specified, including model specifications from tidymodels, in a table format.

Examples

library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() %>% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential", model = lm_mod)

#> Error in eval_spec(spec): argument "spec" is missing, with no default

a <- aim(list(lin = b))

#> Error in aim(list(lin = b)): object 'b' not found