Add Arms to the octomod

Add a new hypothesis arm to the octomod object. The hypothesis collected here are all directed at the core dataset specified by the add_core() function.

add_arm(
  octomod,
  title = NULL,
  f = NULL,
  exposure = NULL,
  pattern = "direct",
  approach,
  ...
)

Arguments

octomod
Object of class octomod

title
Character string to identify this arm

f
Formula showing relationship of outcomes and predictors, and is the essentially the hypothesis.

exposure
Variable(s) that are forced to be maintained in every model as a predictor.

pattern
The variable relationship approach that will be used. The options for the pattern are:

direct will define the relationship as y ~ x

sequential will define the relationship as y ~ x1, y ~ x1 + x2

parallel will define the relationship as y ~ x1, y ~ x2

approach
This describes the analysis plan that should be used. There are several options on which specification to use, anywhere from regression modeling to inferential statistics. The approach applies to the entirety of the growing arm.

A model specified by the parsnip package, which includes the mode and computational engine

A statistical test, such as a t.test, which may require additional parameters. These can be given as additional, unmatched arguments. This option currently supports only hypothesis tests, of class htest.

...
This should reflect the additional parameters that may need to be given to the approach argument, such as paired = TRUE for t.test. The additional parameters must be named to allow them to be passed successfully.

Value

An octomod object with arms attached

Details

This step is the most crucial, in that the the hypothesis and relationship between variables is defined at this stage.

Examples

library(magrittr)

#> Warning: package 'magrittr' was built under R version 4.0.3

om <-
  octomod() %>%
  add_core(mtcars) %>%
  add_arm(
    title = "Horsepower",
    f = disp ~ hp,
    pattern = "direct",
    approach = "t.test",
    paired = TRUE
  )