Create Hypothetical Bullets

A bullet object is a pre-specified hypothesis, with a formula that helps to identify the individual variables and their roles. It also asks for the how the outcome ~ exposure relationship(s) should be defined. The purpose of this is to simplify the creating of a project, where the relationships are defined early, and tested later on.

bullet(f, exposure = NULL, approach = "simple", spec, ...)

Arguments

f
Formula showing relationship of outcomes and predictors

exposure
Variable(s) that are forced to be maintained in every model as a predictor.

approach
The variable relationship approach that will be used. The options are:

plain will define a relationship as y ~ x (however not necessary that x predicts y)

sequential will defines relationship as y ~ x1, y ~ x1 + x2

parallel will define relationships as y ~ x1, y ~ x2

spec
This describes the analysis plan that should be used to help aim the bullet. There are several options on specifications to use, anywhere from regression modeling to inferential statistics. The specification applies to the entire list of bullets being used.

A model specified by the parsnip package, which includes the mode and computational engine

A statistical test, such as a t.test, which may require additional parameters. These can be given as additional, unmatched arguments. This option currently supports only hypothesis tests, of class htest.

...
For extensibility

Value

Object of type bullet that contains the hypothesis and the characteristics of how to put the analysis together.

Examples

library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() %>% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential")

#> Error in eval_spec(spec): argument "spec" is missing, with no default