Equip the octomod With Hypothesis Testing

This function allows for fitting the many arms of the octomod object, equipping them with the correct test methods. For modeling, this heavily leans on the tidymodels approach which allows for unified model specification. For inferential statistics, which do not yet have a unified testing approach, this provides the ability to run these somewhat simpler tests in a cohesive manner.

equip(octomod, which_arms = NULL, ...)

Arguments

octomod
Object of class octomod

which_arms
Vector of names of arms that should be run. Defaults to all arms that have not yet been equipped..

...
To pass additional parameters as needed

Value

An octomod object with arms equipped with the test statistics.

Examples

library(magrittr)
library(parsnip)
lm_mod <- linear_reg() %>% set_engine("lm")

om <-
  octomod() %>%
  core(iris) %>%
  arm(
    title = "t_test",
    plan = Sepal.Length + Sepal.Width ~ Petal.Length,
    pattern = "direct",
    approach = "t.test",
    paired = TRUE
  ) %>%
  arm(
    title = "linear",
    plan = Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length,
    pattern = "sequential",
    approach = lm_mod
  ) %>%
  equip()