Create a Many-Armed Structure to Hold Hypotheses

This creates a central structure that will have many "arms" based on the number of hypotheses that exist for a certain research question. The structure manifests the promise that a single data set can be used to pose multiple hypotheses. These hypothesis should be pre-specified to help structure a research project. At any point, the structure of hypotheses should be agile and malleable.

octomod()

Value

octomod object

Details

The term octomod was used as a play on the concept of how many arms an octopus has, and in this case, it can be used to hold specific models or ideas. This modeling beast can help us tackle complex research projects sensibly and flexibly.